refactor(news-directory): Added News directory to pages and moved related pages accordingly
