+++ client/src/App.tsx
@@ -2,10 +2,10 @@ import { useEffect } from "react"
 import { Route, Routes, useLocation } from "react-router-dom"
 
 import Navbar from "./components/Navbar/Navbar"
-import BreakingNews from "./pages/BreakingNews"
-import NewsPage from "./pages/NewsPage"
-import NewsSubPage from "./pages/NewsSubPage"
-import SportsPage from "./pages/SportsPage"
+import BreakingNews from "./pages/News/BreakingNews"
+import NewsPage from "./pages/News/NewsPage"
+import NewsSubPage from "./pages/News/NewsSubPage"
+import SportsPage from "./pages/News/SportsPage"
 import StaffPage from "./pages/StaffPage"
 import ArticlesTable from "./pages/BusinessPanel/ArticlesTable"
 import ArticleForm from "./pages/BusinessPanel/ArticleForm"
@@ -14,10 +14,10 @@ import UserForm from "./pages/BusinessPanel/UserForm"
 import BusinessSettings from "./pages/BusinessPanel/BusinessSettings"
 import ArticlePage from "./pages/ArticlePage"
 import SubscribePage from "./pages/SubscribePage"
-import LoginPage from "./pages/LoginPage"
-import RegisterPage from "./pages/RegisterPage"
-import ResetPassword from "./pages/ResetPasswordPage"
-import ConfirmResetPassword from "./pages/ConfirmResetPasswordPage"
+import LoginPage from "./pages/Authentication/LoginPage"
+import RegisterPage from "./pages/Authentication/RegisterPage"
+import ResetPassword from "./pages/Authentication/ResetPasswordPage"
+import ConfirmResetPassword from "./pages/Authentication/ConfirmResetPasswordPage"
 import Error from "./pages/Error"
 
 // import ArticleTestData from "./test/ArticleData"
