+++ client/src/pages/News/BreakingNews.tsx
@@ -2,15 +2,15 @@ import axios from "axios"
 import { useState, useEffect } from "react"
 import { useLocation } from "react-router-dom"
 
-import useLoadingNewsPage from "../hooks/useLoadingNewsPage"
+import useLoadingNewsPage from "../../hooks/useLoadingNewsPage"
 
-import { Article } from "../types/interfaces/Article"
+import { Article } from "../../types/interfaces/Article"
 
 
-import ArticleThumbnail from "../components/ArticleThumbnails/ArticleThumbnail"
-import ArticleCarousel from "../components/ArticleCarousel"
-import HeadlineColumn from "../components/BreakingNews/HeadlineColumn"
-import HeadlineBulletPoints from "../components/BreakingNews/HeadlineBulletPoints"
+import ArticleThumbnail from "../../components/ArticleThumbnails/ArticleThumbnail"
+import ArticleCarousel from "../../components/ArticleCarousel"
+import HeadlineColumn from "../../components/BreakingNews/HeadlineColumn"
+import HeadlineBulletPoints from "../../components/BreakingNews/HeadlineBulletPoints"
 
 import "./BreakingNews.css"
